refactor: Common CgroupManager and namespace
