+++ src/Craned/Core/CranedPublicDefs.h
@@ -23,6 +23,7 @@
 #include "PreCompiledHeader.h"
 // Precompiled header comes first
 
+#include "CgroupManager.h"
 #include "CommonPublicDefs.h"
 #include "crane/Network.h"
 #include "crane/OS.h"
@@ -33,7 +34,9 @@ inline constexpr uint64_t kEvSigChldResendMs = 500;
 constexpr uint64_t kCtldClientTimeoutSec = 30;
 constexpr int64_t kCranedRpcTimeoutSeconds = 5;
 
-using EnvMap = std::unordered_map<std::string, std::string>;
+using Common::CgroupInterface;
+using Common::CgroupManager;
+using Common::EnvMap;
 using RegToken = google::protobuf::Timestamp;
 
 enum class CallbackInvokeMode : std::uint8_t { SYNC = 0, ASYNC };
@@ -60,10 +63,97 @@ struct TaskInfoOfUid {
   std::string cgroup_path;
 };
 
+struct Partition {
+  std::unordered_set<std::string> nodes;
+  std::unordered_set<std::string> AllowedAccounts;
+};
+
+struct Config {
+  struct CranedConfig {
+    uint32_t PingIntervalSec;
+    uint32_t CtldTimeoutSec;
+  };
+  CranedConfig CranedConf;
+  struct CranedListenConf {
+    std::string CranedListenAddr;
+    std::string CranedListenPort;
+
+    struct TlsCertsConfig {
+      bool Enabled{false};
+      TlsCertificates TlsCerts;
+      std::string DomainSuffix;
+    };
+
+    TlsCertsConfig TlsConfig;
+
+    std::string UnixSocketListenAddr;
+    std::string UnixSocketForPamListenAddr;
+  };
+  struct ContainerConfig {
+    bool Enabled{false};
+    std::filesystem::path TempDir;
+    std::string RuntimeBin;
+    std::string RuntimeState;
+    std::string RuntimeRun;
+    std::string RuntimeKill;
+    std::string RuntimeDelete;
+  };
+  ContainerConfig Container;
+
+  struct PluginConfig {
+    bool Enabled{false};
+    std::string PlugindSockPath;
+  };
+  PluginConfig Plugin;
+
+  struct SupervisorConfig {
+    std::filesystem::path Path;
+    std::string DebugLevel;
+    std::filesystem::path LogDir;
+  };
+  SupervisorConfig Supervisor;
+
+  CranedListenConf ListenConf;
+  bool CompressedRpc{};
+
+  std::string ControlMachine;
+  std::string CraneCtldForInternalListenPort;
+  std::string CranedDebugLevel;
+  std::string CraneClusterName;
+  uint32_t ConfigCrcVal;
+
+  std::filesystem::path CraneBaseDir;
+  std::filesystem::path CranedLogFile;
+  std::filesystem::path CranedMutexFilePath;
+  std::filesystem::path CranedScriptDir;
+  std::filesystem::path CranedUnixSockPath;
+  std::filesystem::path CranedForPamUnixSockPath;
+
+  bool CranedForeground{};
+
+  std::string Hostname;
+  CranedId CranedIdOfThisNode;
+
+  struct CranedMeta {
+    SystemRelInfo SysInfo;
+    absl::Time CranedStartTime;
+    absl::Time SystemBootTime;
+    std::vector<crane::NetworkInterface> NetworkInterfaces;
+  };
+  CranedMeta CranedMeta;
+
+  std::unordered_map<ipv4_t, std::string> Ipv4ToCranedHostname;
+  std::unordered_map<ipv6_t, std::string, absl::Hash<ipv6_t>>
+      Ipv6ToCranedHostname;
+  std::unordered_map<std::string, std::shared_ptr<ResourceInNode>> CranedRes;
+  std::unordered_map<std::string, Partition> Partitions;
+};
+
+inline Config g_config{};
+
 struct RunTimeStatus {
   std::shared_ptr<spdlog::async_logger> conn_logger;
 };
-using namespace Common;
 
 inline RunTimeStatus g_runtime_status{};
 }  // namespace Craned
