+++ src/Craned/Core/CranedServer.cpp
@@ -338,8 +338,7 @@ grpc::Status CranedServiceImpl::StepStatusChange(
   return Status::OK;
 }
 
-CranedServer::CranedServer(
-    const Common::Config::CranedListenConf &listen_conf) {
+CranedServer::CranedServer(const Config::CranedListenConf &listen_conf) {
   m_service_impl_ = std::make_unique<CranedServiceImpl>();
 
   grpc::ServerBuilder builder;
